requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
algokit-client-generator==2.1.0
algokit-utils==4.0.1
algorand-python==2.7.0
algorand-python-testing==0.4.1
anyio==4.9.0
attrs==25.3.0
black==25.1.0
boolean.py==5.0
CacheControl==0.14.3
cattrs==24.1.3
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
coincurve==21.0.0
coverage==7.8.0
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
distlib==0.3.9
docstring_parser==0.16
ecdsa==0.19.1
filelock==3.18.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.10
idna==3.10
immutabledict==4.2.1
iniconfig==2.1.0
license-expression==30.4.1
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.4.3
mypy==1.15.0
mypy_extensions==1.1.0
networkx==3.4.2
nodeenv==1.9.1
packageurl-python==0.16.0
packaging==24.2
pathspec==0.12.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
pluggy==1.5.0
pre_commit==4.2.0
propcache==0.3.1
puyapy==4.8.0
py-algorand-sdk==2.8.0
py-serializable==2.0.0
pycparser==2.22
pycryptodomex==3.22.0
Pygments==2.19.1
PyNaCl==1.5.0
pyparsing==3.2.3
pytest==8.3.5
pytest-cov==6.1.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
ruff==0.9.10
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
structlog==25.3.0
toml==0.10.2
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.31.1
yarl==1.20.0
